# specs/001-user-authentication-onboarding/contracts/api-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LeaveLab Authentication API",
    "description": "Authentication and user management API for LeaveLab platform",
    "version": "1.0.0",
    "contact": {
      "name": "LeaveLab API Support",
      "email": "api@leavelab.com"
    }
  },
  "servers": [
    {
      "url": "https://leavelab.com/api/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/profile": {
      "get": {
        "summary": "Get current user profile",
        "description": "Retrieve authenticated user's profile information",
        "operationId": "getProfile",
        "tags": ["Profile"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update user profile",
        "description": "Update authenticated user's profile information",
        "operationId": "updateProfile",
        "tags": ["Profile"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "bio": {
            "type": "string",
            "maxLength": 500
          },
          "timezone": {
            "type": "string",
            "example": "Europe/London"
          },
          "language": {
            "type": "string",
            "example": "en-GB"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "email_confirmed": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-08T12:00:00Z"
          },
          "last_sign_in_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-08T14:30:00Z"
          }
        }
      },
      "UserProfile": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "properties": {
              "display_name": {
                "type": "string",
                "example": "John Doe"
              },
              "avatar_url": {
                "type": "string",
                "format": "uri",
                "nullable": true,
                "example": "https://storage.leavelab.com/avatars/user-123.webp"
              },
              "bio": {
                "type": "string",
                "nullable": true,
                "example": "Aspiring digital nomad learning to build AI agencies"
              },
              "timezone": {
                "type": "string",
                "example": "Europe/London"
              },
              "language": {
                "type": "string",
                "example": "en-GB"
              },
              "membership_tier": {
                "type": "string",
                "enum": ["free", "basic", "premium"],
                "example": "free"
              },
              "onboarding_completed": {
                "type": "boolean",
                "example": false
              }
            }
          }
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "example": "AUTH_INVALID_CREDENTIALS"
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "description": "Additional error context"
          }
        },
        "required": ["error"]
      }
    }
  },
...
}